backend/ai_service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi>=0.100.0
starlette>=0.27.0
uvicorn[standard]>=0.20.0
langchain>=0.3.0
langchain-community>=0.0.30
langchain-core>=0.2.0
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
langchain-google-genai>=0.0.10
sentence-transformers>=2.0.0
pdfplumber>=0.9.0
python-dotenv>=1.0.0
diskcache>=5.6.0
pydantic>=2.0.0
supabase>=2.0.0
pgvector>=0.2.0
postgrest>=0.15.0
python-multipart>=0.0.5
redis>=5.0.0
hiredis>=2.2.3
google-generativeai>=0.4.0
torch==2.8.0+cpu
locust>=2.0.0
upstash-search>=0.1.1
tavily-python>=0.7.10
ruff>=0.1.0